+++ generic/__init__.py
@@ -1,1 +1,2 @@
+from .helpers import exec_command
 from .remote_connection import remote_connection