+++ provisioner/__init__.py
@@ -1,2 +1,3 @@
-from .models import CertsInfo, ComponentInfo, Input, Inventory
+from .main import main as provisioner_main
+from .models import CertsInfo, ComponentInfo, Input
 from .provisioner import Provisioner