+++ provisioner/models/component_info.py
@@ -2,7 +2,7 @@ from typing import List
 
 from pydantic import AnyUrl, BaseModel
 
-from provisioner.utils import Component
+from utils import Component
 
 
 class ComponentInfo(BaseModel):